+++ Functions/Public/Restart-PSIISSite.ps1
@@ -9,6 +9,8 @@
     Specify the remote server to run against.
 .PARAMETER Name
     Specify the website name to restart.
+.PARAMETER PassThru
+    If true, the command will return the IIS information.
 .EXAMPLE
     PS> Restart-PSIISSite -ComputerName WebServer01 -Name DefaultSite
 
@@ -32,7 +34,9 @@ Function Restart-PSIISSite() {
         
         [Parameter(ValueFromPipelineByPropertyName)]
         [Alias('Sitename')]
-        [System.String] $Name
+        [System.String] $Name,
+
+        [switch]$PassThru
     )
 
     Begin {}
@@ -73,18 +77,18 @@ Function Restart-PSIISSite() {
             $ScriptBlock = {
                 [CmdletBinding()]
                 Param(
-                    $Site
+                    $Site,
+                    [switch]$PassThru
                 )
                 Import-Module WebAdministration
                 Stop-Website -Name $Site.Name -ErrorAction SilentlyContinue
-                Start-Website -Name $Site.Name -ErrorAction SilentlyContinue
-                Get-PSIISSite -Name $Site.Name -State 'Started'
+                Start-Website -Name $Site.Name -ErrorAction SilentlyContinue -PassThru:$PassThru
             }
             $Site.ComputerName | FOreach-Object {
                 If (IsLocal $_) {
-                    & $ScriptBlock -Site $Site
+                    & $ScriptBlock -Site $Site -PassThru:$PassThru
                 } Else {
-                    Invoke-Command -ComputerName $_ -ScriptBlock $ScriptBlock -ArgumentList $Site | Select-Object -ExcludeProperty RunspaceId
+                    Invoke-Command -ComputerName $_ -ScriptBlock $ScriptBlock -ArgumentList $Site, $PassThru | Select-Object -ExcludeProperty RunspaceId
                 }
             }
         }