+++ Functions/Public/Start-PSIISSite.ps1
@@ -2,9 +2,17 @@
 
 <#
 .SYNOPSIS
-
+    Start an IIS Site.
 .DESCRIPTION
-
+    Start an IIS Site.
+.PARAMETER ComputerName
+    Specify a remote computer to run against.
+.PARAMETER Name
+    Specify the name of the IIS Site to search for.
+.PARAMETER PassThru
+    If true, the command will return the IIS information.
+.EXAMPLE
+    Start-PSIISSite -ComputerName localhost -Name MySite
 .NOTES
     Author:  matthewjdegarmo
     GitHub:  https://github.com/matthewjdegarmo
@@ -22,7 +30,9 @@ Function Start-PSIISSite() {
         
         [Parameter(ValueFromPipelineByPropertyName)]
         [Alias('Sitename')]
-        [System.String] $Name
+        [System.String] $Name,
+
+        [switch]$PassThru
     )
 
     Begin {}
@@ -63,17 +73,17 @@ Function Start-PSIISSite() {
             $ScriptBlock = {
                 [CmdletBinding()]
                 Param(
-                    $Site
+                    $Site,
+                    [switch]$PassThru
                 )
                 Import-Module WebAdministration
-                Start-Website -Name $Site.Name -ErrorAction SilentlyContinue
-                Get-PSIISSite -Name $Site.Name -State 'Started'
+                Start-Website -Name $Site.Name -ErrorAction SilentlyContinue -PassThru:$PassThru
             }
-            $Site.ComputerName | FOreach-Object {
+            $Site.ComputerName | Foreach-Object {
                 If (IsLocal $_) {
-                    & $ScriptBlock -Site $Site
+                    & $ScriptBlock -Site $Site -PassThru:$PassThru
                 } Else {
-                    Invoke-Command -ComputerName $_ -ScriptBlock $ScriptBlock -ArgumentList $Site | Select-Object -ExcludeProperty RunspaceId
+                    Invoke-Command -ComputerName $_ -ScriptBlock $ScriptBlock -ArgumentList $Site, $PassThru | Select-Object -ExcludeProperty RunspaceId
                 }
             }
         }